v1.6.0 #5 Clarified "enable auto color" means "auto apply", and the button changes to "FORCE" Updated project files for new S&M actions Jeffos, the (int) cast in your "don't change without asking SNM_ChunkParserPatcher.h"  fixes a x64 compiler warning

+++ SnM/SNM_ChunkParserPatcher.h
@@ -116,7 +116,7 @@ int ReplaceSubChunk(const char* _keyword, int _depth, int _occurence,
 	if (_keyword && _depth > 0)
 	{
 		WDL_String startToken;
-		startToken.SetFormatted(strlen(_keyword)+1, "<%s", _keyword);
+		startToken.SetFormatted((int)strlen(_keyword)+1, "<%s", _keyword);
 		return ParsePatch(true, SNM_REPLACE_SUB_CHUNK, _depth, _keyword, startToken.Get(), -1, _occurence, -1, (void*)_newSubChunk);
 	}
 	return -1;
